Реализовано обновление статуса пользователя
