+++ src/view/user-profile.js
@@ -1,25 +1,9 @@
 import AbstractView from './abstract.js';
+import {getUserRank} from '../utils/user-profile.js';
 
 const createUserProfileTemplate = (filmCards, userProfileData) => {
-  // Cards with watched flag
-  const watchedFilmsAmount = filmCards.filter((el) => el.isWatched).length;
-  // User profile grade name by default
-  let userGradeName = ``;
 
-  if (watchedFilmsAmount > 0) {
-
-    for (const grade in userProfileData) {
-      if (userProfileData.hasOwnProperty(grade)) {
-        const minFilms = userProfileData[grade].min;
-        const maxFilms = userProfileData[grade].max;
-        userGradeName = (watchedFilmsAmount >= minFilms && watchedFilmsAmount <= maxFilms) ? grade : false;
-
-        if (userGradeName) {
-          break; // If user name is find, break cycle
-        }
-      }
-    }
-  }
+  const userGradeName = getUserRank(filmCards, userProfileData);
 
   return (
     `<section class="header__profile profile">
